support limiting printed bytes
Often I'm interested in looking at headers found in the first bytes of
some data. The data may be quite large, though, so I'd rather not
print the whole thing. It's convenient for pretty_hex to do this for
me, using the complete byte range for the title and printing a trailer
saying how many bytes were skipped.

This is a breaking change because callers may be constructing the struct
exhaustively.